cv.tree {tree}	R Documentation

Cross-validation for Choosing Tree Complexity

Description

Runs a K-fold cross-validation experiment to find the deviance or number of misclassifications as a function of the cost-complexity parameter k.

Usage

cv.tree(object, rand, FUN = prune.tree, K = 10, ...)

Arguments

object
An object of class "tree".

rand
Optionally an integer vector of the length the number of cases used to create object, assigning the cases to different groups for cross-validation.

FUN
The function to do the pruning.

K
The number of folds of the cross-validation.

...
Additional arguments to FUN.

Value

A copy of FUN applied to object, with component dev replaced by the cross-validated results from the sum of the dev components of each fit.

Author(s)

B. D. Ripley

See Also

tree, prune.tree

Examples

data(cpus, package="MASS")
cpus.ltr <- tree(log10(perf) ~ syct + mmin + mmax + cach
     + chmin + chmax, data=cpus)
cv.tree(cpus.ltr, , prune.tree)

[Package tree version 1.0-37 Index]