deviance.tree {tree}	R Documentation

Extract Deviance from a Tree Object

Description

Extract deviance from a tree object.

Usage

## S3 method for class 'tree'
deviance(object, detail = FALSE, ...)

Arguments

object
an object of calls "tree"

detail
logical. If true, returns a vector of deviance contributions from each node.

...
arguments to be passed to or from other methods.

Value

The overall deviance, or a vector of contributions from the cases at each node. The overall deviance is the sum over leaves in the latter case.

[Package tree version 1.0-37 Index]