misclass.tree {tree}	R Documentation

Misclassifications by a Classification Tree

Description

Report the number of mis-classifications made by a classification tree, either overall or at each node.

Usage

misclass.tree(tree, detail = FALSE)

Arguments

tree
Object of class "tree", representing a classification tree.

detail
If false, report overall number of mis-classifications. If true, report the number at each node.

Details

The quantities returned are weighted by the observational weights if these are supplied in the construction of tree.

Value

Either the overall number of misclassifications or the number for each node.

Author(s)

B. D. Ripley

See Also

tree

Examples

ir.tr <- tree(Species ~., iris)
misclass.tree(ir.tr)
misclass.tree(ir.tr, detail=TRUE)

[Package tree version 1.0-37 Index]