tree.screens {tree}	R Documentation

Split Screen for Plotting Trees

Description

Splits the screen in a way suitable for using tile.tree.

Usage

tree.screens(figs, screen.arg = 0, ...)

Arguments

figs
A specification of the split of the screen. See split.screen for the allowed forms.

screen.arg
the screen to divide, by default the whole display area.

...
plot parameters to be passed to par.

Value

A vector of screen numbers for the newly-created screens.

Author(s)

B. D. Ripley

See Also

tile.tree, split.screen

Examples

data(fgl, package="MASS")
fgl.tr <- tree(type ~ ., fgl)
summary(fgl.tr)
plot(fgl.tr);  text(fgl.tr, all=TRUE, cex=0.5)
fgl.tr1 <- snip.tree(fgl.tr, node=c(108, 31, 26))
tree.screens()
plot(fgl.tr1)
tile.tree(fgl.tr1, fgl$type)
close.screen(all = TRUE)

[Package tree version 1.0-37 Index]